Make the content more customizable
Use a different environment variable extension for Markdown that allows
using conditionals:
http://andrewjrobinson.github.io/python-md-environment/
With this extension the syntax is a bit different and we need to replace
instances of ${ENV_VAR} with \env{ENV_VAR}.

To make it possible to customize the system links on the home page,
extend the readthedocs theme with a custom main.html that contains the
links. Because main.html is a template, we can fetch data from the extra
dict in mkdocs.yml.

Make mkdocs.yml into a Jinja2 template. Add a script that calls yasha to
customize the template based on environment variables. The environment
variables must be set or the script returns with a non-zero exit code.
Add mkdocs.yml into .gitignore as we don't want a compiled version of
the config to end up in version control.

Modify the Dockerfile so that it first runs the make_config.sh script to
create configuration before starting mkdocs serve.