Use Nginx for hosting site
Instead of using the mkdocs preview server in the Dockerfile, use Nginx
instead. The web server is installed on a centos:7 base image and
configuration via environment variables is moved from run time to build
time. Update README.md accordingly and also add instructions for hosting
the guide on OpenShift.