Project documentation
Add documentation for:
 - Creating a project that is mapped to a CSC computing project
 - Project quotas
 - Sharing projects with other users

Reorganize the "Getting started" documentation a bit as there is now a
separate page for project management documentation.

Also add a work-in-progress page on billing that is hidden by default.
This is added to demonstrate how to add hidden pages. More information
can be filled in once we have all the details of billing.

Also add a helper script set_env_rahti.sh that can be used to set the
environment variables to what they would be set to in Rahti. This is
useful during development when you want to preview what the Rahti user
guide will actually look like once it is live.

Because of the added content, make subheadings visible in the sidebar.
These were previously hidden as most documentation pages were quite
short.

Some of the content under "Technical maturity" has also been updated for
clarity.