Fix Nginx configuration for OpenShift
Fix Nginx configuration so that it redirects links like "access" to
"access/". At the same time create a separate file for the config that
is copied into the Docker container image instead of using sed to modify
the config file. This makes it much easier to make changes to Nginx
config in the future.

NB: This breaks the current Dockerfile for running locally on one's
laptop, but it's more important that this works in OpenShift. We can fix
the local case next, but the priority should be to fix what's running in
OpenShift.