Make Nginx redirects relative
Instead of using absolute redirects that also include the scheme, use
relative redirects instead. This way Nginx won't try to redirect the
user to the http scheme from https when running in OpenShift and issuing
301s.

Update the documentation for running Nginx locally to use port 80
instead of 8000 so that these new relative redirects also work when
testing Docker changes locally. The downside is that port 80 must be
available locally, but this is generally the case and if a user runs
Docker locally they should also have root locally to be able to bind to
port 80.